Move FunctionalOperator to execute code
