+++ java/execute/src/com/ibm/streamsx/topology/spi/operators/Pipe.java
@@ -8,7 +8,6 @@ import com.ibm.streams.operator.model.OutputPorts;
 import com.ibm.streams.operator.model.SharedLoader;
 import com.ibm.streamsx.topology.function.Consumer;
 import com.ibm.streamsx.topology.internal.functional.ops.AbstractPipe;
-import com.ibm.streamsx.topology.spi.runtime.FunctionalOperator;
 
 /**
  * A functional pipe operator.
