+++ java/execute/src/com/ibm/streamsx/topology/spi/operators/Primitive.java
@@ -4,7 +4,6 @@ import com.ibm.streams.operator.OperatorContext;
 import com.ibm.streams.operator.model.SharedLoader;
 import com.ibm.streamsx.topology.function.ObjIntConsumer;
 import com.ibm.streamsx.topology.internal.functional.ops.AbstractPrimitive;
-import com.ibm.streamsx.topology.spi.runtime.FunctionalOperator;
 
 /**
  * A functional operator.
